\name{Blaetter}
\alias{Blaetter}
\title{ die Indices der Blaetter eines Graphen }
\arguments{
\item{Adj[1:n,1:n]}{Adjazenzmatrix  passend zu den folgenden Knoten Infos}
}
\value{BlattInd[1:b]          Indizes, so dass Adj[,BlattInd] sind die Bletter des Graphen}
\description{
 die Indices der Blaetter eines Graphen }
\author{
}
