\name{CornerPoints}
\alias{CornerPoints}
\title{ Ergänzt Punkte an den Ecken, die einen Datensatz toroid machen. }
\usage{
CornerPoints(X, Y, Cls)
}
\arguments{
\item{X, Y: Vektoren mit koordinaten der Datenpunkte}{}
\item{Cls: Vektor mit der Klasse der Datenpunkte, falls nicht gegeben, sind alle Daten Cls = 1.}{}
}
\value{XplusCorner, YplusCorner, ClsPlusCorner: Vektoren X, Y und CLS, an deren Ende jeweils die zusätzlichen Punkte, bzw. deren Klasse angefügt wurde.}
\description{
 Ergänzt Punkte an den Ecken, die einen Datensatz toroid machen. }
\author{
}
