\name{DelaunayGraphMatrix}
\alias{DelaunayGraphMatrix}
\title{ Calculates the Adjacency Martix for a Delaunay-Graph}
\usage{
DelaunayGraphMatrix(X, Y, Tiling)
}
\arguments{
\item{X[1:n],}{Y[1:n]         point coordinates for which a planar Delaunay graph is constucted POINTS NEED NOT BE UNIQUE! however, there is an edge in the Delaunay between duplicate points!}
\item{PlotIt}{if TRUE Delaunay graph is plotted}
}
\value{a list of}
\description{
 Calculates the Adjacency Martix for a Delaunay-Graph}
\author{
}
