\name{InnerVsTransDelaunayKanten}
\alias{InnerVsTransDelaunayKanten}
\title{ calculates which edges of the Delaunay-Graph lie in one class (inner) and which lie in two classes (trans).}
\arguments{
\item{Delaunay}{the adjacency matrix of the DelaunayGraph}
\item{Cls}{The classes of the BestMatches}
}
\value{TransClsInd: vector containing the indices of those points which have edges to points in different classes}
\description{
 calculates which edges of the Delaunay-Graph lie in one class (inner) and which lie in two classes (trans).}
