\name{PlotDelaunay4BestMatches}
\alias{PlotDelaunay4BestMatches}
\title{ Zeichnen des Delaunay Graphen, Punkte ggf. durch Cls gefaerbt}
\arguments{
\item{BestMatches(1:d,1:3)}{Bestmatches}
\item{MatrixOrSize[2]}{Default 50,82; A vector of length 2, containing the number of lines and columns of the SOM corresponding to the BestMatches}
\item{Cls(1:d)}{Classes of bestmatches  or []}
\item{IsToroid}{==1 (default) => Randkorrigiertes Universum}
\item{USES}{Delaunay4BestMatches(BestMatches,MatrixOrSize,IsToroid);}
}
\description{
 Zeichnen des Delaunay Graphen, Punkte ggf. durch Cls gefaerbt
}
\author{
Author: RG Nov.2014
}