\name{PlotVoronoiPoliticalMap}
\alias{PlotVoronoiPoliticalMap}
\title{ Zeichnen der Voronoizellen eingefaerbt mit den ClsColors default mit Border correction, so dass keine offenen ungefaebten Randzellen zu sehen sind}
\arguments{
\item{[X(1:d),Y(1:d)]}{Punktkoordinaten}
\item{Cls(1:d)}{Classes of bestmatches  or []}
\item{BorderCorrection}{==1 (default) => Randkorrigiertes Universum}
\item{showpoints}{TRUE, Punkte werden angezeigt}
\item{uses}{CornerPoints(X,Y,Cls) DefaultColors4Cls(Cls) package deldir}
}
\description{
 Zeichnen der Voronoizellen eingefaerbt mit den ClsColors default mit Border correction, so dass keine offenen ungefaebten Randzellen zu sehen sind}

