\name{RkugelGraph}
\alias{RkugelGraph}
\title{ Berechnung des Rkugel Graphen  }
\arguments{
\item{DataOrDists(1:n,1:n)}{array of data: n cases in rows, d variables in columns, matrix is not symmetric or distance matrix, in this case matrix has to be symmetric}
\item{R}{Kugelradius}
\item{PlotIt}{}
\item{Points}{}
}
\value{RKugelGraphAdjMatrix                   Rkugel Graphen}
\description{
 Berechnung des Rkugel Graphen  }
\author{
}
