\name{TransClsDelaunay}
\alias{TransClsDelaunay}
\title{ calculates all points with Delaunay-Edges crossing into different classes.}
\arguments{
\item{TRI}{A n by 3 matrix containing the three points forming a Delaunay-Triangle in rows}
\item{Cls}{Vector containing the classes of bestmatches.}
}
\value{List with
  \item{TransCLSPointInd}{}
  \item{TransClsTRI}{}
  \item{TransAInd}{}
  \item{TransBInd}{}
  \item{InnerPointInd}{}
}
\description{
 calculates all points with Delaunay-Edges crossing into different classes.}

