\name{UGraph}
\alias{UGraph}
\title{ Calculates the UGraph }
\arguments{
\item{Data}{The original data, as a n-by-d-matrix. N data-points in a d-dimensional data-space are the rows.}
\item{Positions}{n-by2-Matrix, containing the coordinates in the 2-dimensional output-space. The same as BestMatches[,2:3]}
}
\value{A list of:
  \item{Uheights}{}
  \item{VoronoiMeanPts}{}
  \item{UheightBorder}{}
  \item{BorderPt1}{}
  \item{BorderPt2}{}
  \item{Aind}{}
  \item{Bind}{}
}
\description{
 Calculates the UGraph 
}

