\name{UntileGraph4BM}
\alias{UntileGraph4BM}
\title{ Given the adjacency matrix of a tiled graph and the tiled bestmatches, this returns the adjacency matrix of the untiled graph.}
\arguments{
\item{TiledGraph: The adjacency matrix of the tiled graph}{}
\item{TiledBestMatches: The tiled BestMatches}{}
\item{Lines, Columns: The number of lines and columns of the original SOM}{}
}
\value{UntiledGraph: the adjacency matrix of the untiled graph.}
\description{
 Given the adjacency matrix of a tiled graph and the tiled bestmatches, this returns the adjacency matrix of the untiled graph.}
\author{
}
